Move LegacyAuthentication concern
